feature(btc-server): add feature flags wallet_state_sync and conflicting_input

+++ bin/btc-server/src/signer.rs
@@ -227,7 +227,10 @@ where
         // }
 
         // Validate PSBT
-        validate_psbt(psbt, ROUND1, self.min_signers, &self.db)?;
+        if cfg!(feature = "conflicting_input") {
+            validate_psbt(psbt, ROUND1, self.min_signers, &self.db)?;
+        }
+
         let num_inputs = psbt.inputs.len();
 
         let key_package =
@@ -266,7 +269,9 @@ where
             self.db.get_key_package()?.ok_or(SigningRound2Error::MissingKeyPackage)?;
 
         // Validate PSBT
-        validate_psbt(psbt, ROUND1_TRANSITION, self.min_signers, &self.db)?;
+        if cfg!(feature = "conflicting_input") {
+            validate_psbt(psbt, ROUND1, self.min_signers, &self.db)?;
+        }
 
         let tx = psbt.clone().extract_tx()?;
         let num_inputs = tx.input.len();
@@ -344,9 +349,11 @@ where
         }
         let pending_pegout_ids = pending_pegouts.iter().map(|p| p.id).collect::<Vec<PegoutId>>();
 
-        self.add_tracked_tx(tx.clone(), &pending_pegouts, SystemTime::now()).await?;
-        self.db.remove_pending_pegout(&pending_pegout_ids)?;
-        self.db.flush()?;
+        if cfg!(feature = "conflicting_input") {
+            self.add_tracked_tx(tx.clone(), &pending_pegouts, SystemTime::now()).await?;
+            self.db.remove_pending_pegout(&pending_pegout_ids)?;
+            self.db.flush()?;
+        }
 
         // Clear the signing nonces
         // This finalizes the signing session
